third_party/WebKit/LayoutTests/platform/mac/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/redirect-credentials-base-https-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: https://127.0.0.1:8443/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Flocalhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttps%3A%2F%2F127.0.0.1%3A8443&mode=cors&credentials=omit&method=GET should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
